Update shoppingQueries.sql
Added carthistory table.


documentation: when user has paid for his items in the cart, all the product id,username, total price and date '(add date in Cart.jsp using expression java.util.Date and get the date using parameter when user click pay so that the purchase will use date as a reference for cart history)' will be inserted in the carthistory table. 
Also after inserting, the item in the cart needs to be deleted in the database table cart.

so after paying user should not see any item in his cart rather it will be in the carthistory section of the page.